+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-CHANGELOG for 2.0.x
-===================
+CHANGELOG for 2.x
+=================
 ## 2.1.0
 * [New] Sort items of a collection. See feature #67.
 * [New] Support to extends TwigRenderizer with tags. See PR #65.
@@ -12,6 +12,7 @@ CHANGELOG for 2.0.x
 * [Improved] Improved permalinks customizations. See PR #64.
 * [Improved] Eliminated unnecessary calls to `setItem` method of `SiteAttribute` class in `ContentManager` class.
 * [Improved] The methods for managing the collection of plugins in `PluginManager` class have been extracted to a separated class called `PluginCollection`.
+* [Improved] The methods for managing the collection of collection-item in `CollectionManager` class have been extracted to a separated class called `ColectionItemCollection`.
 * [Fixed] Fixed the path available at `page.path` variable. Prior to this version, this variable contains the relative path to `src/content/` but with the filename extension changed by the Converter. Now, the original filename extension isn't altered. A new path snapshot has been created in `ItemIterface`.
 
 ## 2.0.2 (2016-01-16)